+++ tests/integ/test_tf.py
@@ -20,8 +20,10 @@ import pytest
 from sagemaker.tensorflow import TensorFlow
 from tests.integ import DATA_DIR, TRAINING_DEFAULT_TIMEOUT_MINUTES
 from tests.integ.timeout import timeout_and_delete_endpoint_by_name, timeout
+from tests.integ.vpc_utils import get_or_create_subnet_and_security_group
 
 DATA_PATH = os.path.join(DATA_DIR, 'iris', 'data')
+VPC_NAME = 'training-job-test'
 
 
 @pytest.mark.continuous_testing
@@ -90,6 +92,8 @@ def test_tf_async(sagemaker_session):
 def test_failed_tf_training(sagemaker_session, tf_full_version):
     with timeout(minutes=TRAINING_DEFAULT_TIMEOUT_MINUTES):
         script_path = os.path.join(DATA_DIR, 'iris', 'failure_script.py')
+        ec2_client = sagemaker_session.boto_session.client('ec2')
+        subnet, security_group_id = get_or_create_subnet_and_security_group(ec2_client, VPC_NAME)
         estimator = TensorFlow(entry_point=script_path,
                                role='SageMakerRole',
                                framework_version=tf_full_version,
@@ -98,10 +102,17 @@ def test_failed_tf_training(sagemaker_session, tf_full_version):
                                hyperparameters={'input_tensor_name': 'inputs'},
                                train_instance_count=1,
                                train_instance_type='ml.c4.xlarge',
-                               sagemaker_session=sagemaker_session)
+                               sagemaker_session=sagemaker_session,
+                               subnets=[subnet],
+                               security_group_ids=[security_group_id])
 
         inputs = estimator.sagemaker_session.upload_data(path=DATA_PATH, key_prefix='integ-test-data/tf-failure')
 
         with pytest.raises(ValueError) as e:
             estimator.fit(inputs)
         assert 'This failure is expected' in str(e.value)
+
+        job_desc = estimator.sagemaker_session.sagemaker_client.describe_training_job(
+            TrainingJobName=estimator.latest_training_job.name)
+        assert [subnet] == job_desc['VpcConfig']['Subnets']
+        assert [security_group_id] == job_desc['VpcConfig']['SecurityGroupIds']